+++ src/ReactNativeWebView.bs.js
@@ -1,7 +1,7 @@
 'use strict';
 
-var Event$ReactNative = require("reason-react-native/src/apis/Event.bs.js");
-var NativeElement$ReactNative = require("reason-react-native/src/elements/NativeElement.bs.js");
+var Event$ReactNative = require("rescript-react-native/src/apis/Event.bs.js");
+var NativeElement$ReactNative = require("rescript-react-native/src/elements/NativeElement.bs.js");
 var ReactNativeWebView_UnionCallback = require("./ReactNativeWebView_UnionCallback.bs.js");
 
 var Source = {};