+++ src/ReactNativeWebView_UnionCallback.bs.js
@@ -1,8 +1,8 @@
 'use strict';
 
 var Curry = require("bs-platform/lib/js/curry.js");
-var Event$ReactNative = require("reason-react-native/src/apis/Event.bs.js");
-var NativeElement$ReactNative = require("reason-react-native/src/elements/NativeElement.bs.js");
+var Event$ReactNative = require("rescript-react-native/src/apis/Event.bs.js");
+var NativeElement$ReactNative = require("rescript-react-native/src/elements/NativeElement.bs.js");
 
 function Make(funarg) {
   Event$ReactNative.SyntheticEvent({});